BCM 2712 code merge and update

+++ include/linux/dma-heap.h
@@ -23,8 +23,8 @@ struct dma_heap;
 struct dma_heap_ops {
 	struct dma_buf *(*allocate)(struct dma_heap *heap,
 				    unsigned long len,
-				    u32 fd_flags,
-				    u64 heap_flags);
+				    unsigned long fd_flags,
+				    unsigned long heap_flags);
 };
 
 /**
